refactor: consolidate test helpers to global directory
  - Move all test helpers from src/core/__tests__/__helpers__ to src/__tests__/__helpers__ for project-wide accessibility
  - Update all import paths throughout test files to use consolidated helpers
  - Replace integration test with typia validator instead of manual validation logic
  - Update package.json typia:generate paths to reflect new directory structure
  - Fix TypeScript compilation errors with non-null assertions for test data

  This consolidation improves test infrastructure organization and ensures all test utilities are available project-wide.
