+++ package.json
@@ -28,8 +28,8 @@
     "test:coverage": "npm run typia:generate && vitest run --coverage",
     "test:emulator": "vitest run --config vitest.emulator.config.ts",
     "test:emulator:watch": "vitest --config vitest.emulator.config.ts",
-    "typia:generate": "typia generate --input src/core/__tests__/__helpers__/typia-validators --output src/core/__tests__/__helpers__/typia-validators/__generated__ && npm run typia:postprocess",
-    "typia:postprocess": "find src/core/__tests__/__helpers__/typia-validators/__generated__ -name '*.ts' -exec sed -i '/^import.*typia.*from.*[\"\\x27]typia[\"\\x27];*$/d' {} + || true",
+    "typia:generate": "typia generate --input src/__tests__/__helpers__/typia-validators --output src/__tests__/__helpers__/typia-validators/__generated__ && npm run typia:postprocess",
+    "typia:postprocess": "find src/__tests__/__helpers__/typia-validators/__generated__ -name '*.ts' -exec sed -i '/^import.*typia.*from.*[\"\\x27]typia[\"\\x27];*$/d' {} + || true",
     "lint": "eslint \"src/**/*.ts\" \"*.{ts,js,mjs}\"",
     "lint:fix": "eslint \"src/**/*.ts\" \"*.{ts,js,mjs}\" --fix",
     "typecheck": "npm run typia:generate && tsc --noEmit",
