+++ src/core/__tests__/unit/collection-group.unit.test.ts
@@ -4,14 +4,14 @@ import { Query } from '../../query'
 import { serializeFirestoreTypes } from '../../../utils/firestore-converter'
 import { validateData } from '../../../utils/validator'
 import type { FirestoreTypedOptionsProvider } from '../../../types/firestore-typed.types'
-import type { TestPostEntity } from '../__helpers__/test-entities.helper'
+import type { TestPostEntity } from '../../../__tests__/__helpers__/test-entities.helper'
 
 // Mock dependencies
 vi.mock('../../../utils/firestore-converter')
 vi.mock('../../../utils/validator')
 
 vi.mock('firebase-admin/firestore', async () => {
-  const mockHelper = await import('../__helpers__/firebase-mock.helper')
+  const mockHelper = await import('../../../__tests__/__helpers__/firebase-mock.helper')
   return mockHelper.createFirebaseAdminMock()
 })
